requirements.txt
# GCP
google-cloud-bigquery[pandas]>=3.25
google-auth>=2.29

# Data
pandas>=2.2
pyarrow>=16.0

# ML tracking
mlflow>=2.13

# Orchestration
dagster>=1.7
dagster-webserver>=1.7

# dbt
dbt-bigquery>=1.8

# API
fastapi>=0.111
uvicorn>=0.30

# Dev
pytest>=8.2
python-dotenv>=1.0
